Add files via upload
Zero-Or-All is an activity where users learn about 0s and 1s. When we click, 'Off the lights', it turns the background white; which represents 0, or no input. Similarily, when we click, 'On the lights', it turns the background red; which represents 1, or input. When we click the 'Quit' button, it quits the program, as what better way is there to Alt+F4 than using a quit button?